requirements.txt
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.0
pytorch-grad-cam>=1.5.0
quantus>=0.5.0
scikit-image>=0.21.0
scipy>=1.10.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
datasets>=2.14.0
